+++ src/vars/error.rs
@@ -1,5 +1,5 @@
 use derive_more::From;
-use orion_error::{ErrorCode, StructError, UvsReason};
+use orion_error::{DomainReason, ErrorCode, StructError, UvsReason};
 use serde_derive::Serialize;
 use thiserror::Error;
 #[derive(Clone, Debug, Serialize, PartialEq, Error, From)]
@@ -11,6 +11,8 @@ pub enum VarsReason {
     #[error("{0}")]
     Uvs(UvsReason),
 }
+
+impl DomainReason for VarsReason {}
 
 impl ErrorCode for VarsReason {
     fn error_code(&self) -> i32 {